refactor: into render shortcuts
