+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/esp/modes/Esp2DMode.kt
@@ -21,18 +21,17 @@ package net.ccbluex.liquidbounce.features.module.modules.render.esp.modes
 import net.ccbluex.liquidbounce.event.events.OverlayRenderEvent
 import net.ccbluex.liquidbounce.event.handler
 import net.ccbluex.liquidbounce.features.module.modules.render.esp.ModuleESP.getColor
-import net.ccbluex.liquidbounce.injection.mixins.minecraft.gui.MixinDrawContextAccessor
+import net.ccbluex.liquidbounce.render.drawHorizontalLine
+import net.ccbluex.liquidbounce.render.drawVerticalLine
 import net.ccbluex.liquidbounce.render.engine.type.Color4b
+import net.ccbluex.liquidbounce.render.fill
 import net.ccbluex.liquidbounce.render.renderEnvironmentForGUI
 import net.ccbluex.liquidbounce.utils.entity.RenderedEntities
 import net.ccbluex.liquidbounce.utils.entity.interpolateCurrentPosition
 import net.ccbluex.liquidbounce.utils.render.WorldToScreen
 import net.minecraft.client.gui.DrawContext
-import net.minecraft.client.render.RenderLayer
-import net.minecraft.client.render.VertexConsumer
 import net.minecraft.util.math.Box
 import net.minecraft.util.math.Vec3d
-import org.joml.Matrix4f
 
 object Esp2DMode : EspMode("2D") {
 
@@ -95,18 +94,18 @@ object Esp2DMode : EspMode("2D") {
                         if (outline) {
 
                             if (border) {
-                                horizontalLine(0.0f, rectWidth, 0.0f, 1.5f, black)
-                                verticalLine(0.0f, 0.0f, rectHeight, 1.5f, black)
-                                horizontalLine(0.0f, rectWidth, rectHeight, 1.5f, black)
-                                verticalLine(rectWidth, 0.0f, rectHeight + 1.5f, 1.5f, black)
+                                drawHorizontalLine(0.0f, rectWidth, 0.0f, 1.5f, black)
+                                drawVerticalLine(0.0f, 0.0f, rectHeight, 1.5f, black)
+                                drawHorizontalLine(0.0f, rectWidth, rectHeight, 1.5f, black)
+                                drawVerticalLine(rectWidth, 0.0f, rectHeight + 1.5f, 1.5f, black)
 
                                 translate(0.5f, 0.5f, 0.0f)
                             }
 
-                            horizontalLine(0.0f, rectWidth, 0.0f, 0.5f, outlineColor.toARGB())
-                            horizontalLine(0.0f, rectWidth, rectHeight, 0.5f, outlineColor.toARGB())
-                            verticalLine(0.0f, 0.0f, rectHeight, 0.5f, outlineColor.toARGB())
-                            verticalLine(rectWidth, 0.0f, rectHeight + 0.5f, 0.5f, outlineColor.toARGB())
+                            drawHorizontalLine(0.0f, rectWidth, 0.0f, 0.5f, outlineColor.toARGB())
+                            drawHorizontalLine(0.0f, rectWidth, rectHeight, 0.5f, outlineColor.toARGB())
+                            drawVerticalLine(0.0f, 0.0f, rectHeight, 0.5f, outlineColor.toARGB())
+                            drawVerticalLine(rectWidth, 0.0f, rectHeight + 0.5f, 0.5f, outlineColor.toARGB())
 
                             if (border) {
                                 translate(-0.5f, -0.5f, 0.0f)
@@ -120,8 +119,8 @@ object Esp2DMode : EspMode("2D") {
 
                             translate(-3.0f, 0.0f, 0.0f)
 
-                            if (border) verticalLine(0.0f, 0.0f, rectHeight + 1.5f, 1.5f, black)
-                            verticalLine(0.5f, rectHeight + 1f, rectHeight - healthHeight + 0.5f, 0.5f, healthColor.toARGB())
+                            if (border) drawVerticalLine(0.0f, 0.0f, rectHeight + 1.5f, 1.5f, black)
+                            drawVerticalLine(0.5f, rectHeight + 1f, rectHeight - healthHeight + 0.5f, 0.5f, healthColor.toARGB())
                         }
                     }
                 }
@@ -131,37 +130,3 @@ object Esp2DMode : EspMode("2D") {
 
 }
 
-
-private fun DrawContext.fill(x1: Float, y1: Float, x2: Float, y2: Float, z: Float, color: Int) {
-    val layer = RenderLayer.getGui()
-    var x1 = x1
-    var y1 = y1
-    var x2 = x2
-    var y2 = y2
-    val matrix4f: Matrix4f? = this.matrices.peek().getPositionMatrix()
-    if (x1 < x2) {
-        val i = x1
-        x1 = x2
-        x2 = i
-    }
-
-    if (y1 < y2) {
-        val i = y1
-        y1 = y2
-        y2 = i
-    }
-
-    val vertexConsumer: VertexConsumer = (this as MixinDrawContextAccessor).vertexConsumers.getBuffer(layer)
-    vertexConsumer.vertex(matrix4f, x1, y1, z).color(color)
-    vertexConsumer.vertex(matrix4f, x1, y2, z).color(color)
-    vertexConsumer.vertex(matrix4f, x2, y2, z).color(color)
-    vertexConsumer.vertex(matrix4f, x2, y1, z).color(color)
-}
-
-private fun DrawContext.horizontalLine(x1: Float, x2: Float, y: Float, thickness: Float, color: Int) {
-    this.fill(x1, y, x2, y + thickness, 0f, color)
-}
-
-private fun DrawContext.verticalLine(x: Float, y1: Float, y2: Float, thickness: Float, color: Int) {
-    this.fill(x, y1, x + thickness, y2, 0f, color)
-}
